Rename to `Url::slug()`
