+++ src/Cms/PageActions.php
@@ -153,7 +153,7 @@ trait PageActions
 		string|null $languageCode = null
 	): static {
 		// always sanitize the slug
-		$slug = App::slug($slug);
+		$slug = Url::slug($slug);
 
 		// in multi-language installations the slug for the non-default
 		// languages is stored in the text file. The changeSlugForLanguage
@@ -443,7 +443,7 @@ trait PageActions
 		$files       = $options['files']     ?? false;
 
 		// clean up the slug
-		$slug = App::slug($slug);
+		$slug = Url::slug($slug);
 
 		if ($parentModel->findPageOrDraft($slug)) {
 			throw new DuplicateException([
@@ -495,7 +495,7 @@ trait PageActions
 	public static function create(array $props): Page
 	{
 		// clean up the slug
-		$props['slug']      = App::slug($props['slug'] ?? $props['content']['title'] ?? null);
+		$props['slug']      = Url::slug($props['slug'] ?? $props['content']['title'] ?? null);
 		$props['template']  = $props['model'] = strtolower($props['template'] ?? 'default');
 		$props['isDraft'] ??= $props['draft'] ?? true;
 
@@ -685,7 +685,7 @@ trait PageActions
 	public function duplicate(string|null $slug = null, array $options = []): static
 	{
 		// create the slug for the duplicate
-		$slug = App::slug($slug ?? $this->slug() . '-' . App::slug(I18n::translate('page.duplicate.appendix')));
+		$slug = Url::slug($slug ?? $this->slug() . '-' . Url::slug(I18n::translate('page.duplicate.appendix')));
 
 		$arguments = [
 			'originalPage' => $this,
