+++ tests/Cms/App/AppTest.php
@@ -836,42 +836,6 @@ class AppTest extends TestCase
 		$this->assertSame(['Vary' => 'Cookie', 'Cache-Control' => 'custom'], $app->response()->headers());
 	}
 
-	/**
-	 * @covers ::slug
-	 */
-	public function testSlug()
-	{
-		// default length
-		$this->assertSame(
-			'this-is-a-very-long-sentence-that-should-be-used-to-test-the-app-slug-method-by-default-it-should-limit-the-slug-to-255-characters-so-let-s-see-how-well-this-works-for-this-new-method-would-be-a-pity-if-at-one-point-this-breaks-would-be-a-pity-if-at-one-p',
-			App::slug('This is a very long sentence that should be used to test the App::slug() method. By default it should limit the slug to 255 characters. So let\'s see how well this works for this new method. Would be a pity if at one point this breaks. Would be a pity if at one point this breaks. Would be a pity if at one point this breaks.')
-		);
-
-		// custom length
-		$app = new App([
-			'options' => [
-				'slugs.maxlength' => 40
-			]
-		]);
-
-		$this->assertSame(
-			'this-is-a-very-long-sentence-that-should',
-			App::slug('This is a very long sentence that should be used to test the `slugs.maxlength` option. This should be cut already after 40 characters.')
-		);
-
-		// no length restriction
-		$app = new App([
-			'options' => [
-				'slugs.maxlength' => false
-			]
-		]);
-
-		$this->assertSame(
-			'this-is-a-very-long-sentence-that-should-be-used-to-test-the-app-slug-method-by-default-it-should-limit-the-slug-to-255-characters-but-we-can-disable-the-shortening-so-let-s-see-how-well-this-works-for-this-new-method-would-be-a-pity-if-at-one-point-this-breaks-would-be-a-pity-if-at-one-point-this-breaks-would-be-a-pity-if-at-one-point-this-breaks',
-			App::slug('This is a very long sentence that should be used to test the App::slug() method. By default it should limit the slug to 255 characters, but we can disable the shortening. So let\'s see how well this works for this new method. Would be a pity if at one point this breaks. Would be a pity if at one point this breaks. Would be a pity if at one point this breaks.')
-		);
-	}
-
 	public function testInstance()
 	{
 		App::destroy();
