+++ tests/Cms/Facades/UrlTest.php
@@ -4,6 +4,9 @@ namespace Kirby\Cms;
 
 use Kirby\Filesystem\F;
 
+/**
+ * @coversDefaultClass \Kirby\Cms\Url
+ */
 class UrlTest extends TestCase
 {
 	public const TMP = KIRBY_TMP_DIR . '/Cms.Url';
@@ -25,6 +28,42 @@ class UrlTest extends TestCase
 	public function testHome()
 	{
 		$this->assertSame('https://getkirby.com', Url::home());
+	}
+
+	/**
+	 * @covers ::slug
+	 */
+	public function testSlug()
+	{
+		// default length
+		$this->assertSame(
+			'this-is-a-very-long-sentence-that-should-be-used-to-test-the-url-slug-method-by-default-it-should-limit-the-slug-to-255-characters-so-let-s-see-how-well-this-works-for-this-new-method-would-be-a-pity-if-at-one-point-this-breaks-would-be-a-pity-if-at-one-p',
+			Url::slug('This is a very long sentence that should be used to test the Url::slug() method. By default it should limit the slug to 255 characters. So let\'s see how well this works for this new method. Would be a pity if at one point this breaks. Would be a pity if at one point this breaks. Would be a pity if at one point this breaks.')
+		);
+
+		// custom length
+		$app = new App([
+			'options' => [
+				'slugs.maxlength' => 40
+			]
+		]);
+
+		$this->assertSame(
+			'this-is-a-very-long-sentence-that-should',
+			Url::slug('This is a very long sentence that should be used to test the `slugs.maxlength` option. This should be cut already after 40 characters.')
+		);
+
+		// no length restriction
+		$app = new App([
+			'options' => [
+				'slugs.maxlength' => false
+			]
+		]);
+
+		$this->assertSame(
+			'this-is-a-very-long-sentence-that-should-be-used-to-test-the-url-slug-method-by-default-it-should-limit-the-slug-to-255-characters-but-we-can-disable-the-shortening-so-let-s-see-how-well-this-works-for-this-new-method-would-be-a-pity-if-at-one-point-this-breaks-would-be-a-pity-if-at-one-point-this-breaks-would-be-a-pity-if-at-one-point-this-breaks',
+			Url::slug('This is a very long sentence that should be used to test the Url::slug() method. By default it should limit the slug to 255 characters, but we can disable the shortening. So let\'s see how well this works for this new method. Would be a pity if at one point this breaks. Would be a pity if at one point this breaks. Would be a pity if at one point this breaks.')
+		);
 	}
 
 	public function testTo()
